% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acr_subgroups_paper_colors.R
\name{acr_subgroups_paper_colors}
\alias{acr_subgroups_paper_colors}
\title{Retrieve Colors for ACR Subgroups and Other Variables}
\usage{
acr_subgroups_paper_colors(
  sg1_var_name = "1",
  sg2_var_name = "2",
  sg3_var_name = "3",
  sg4_var_name = "4",
  fem_var_name = "Female",
  mas_var_name = "Male",
  case_var_name = "case",
  control_var_name = "control"
)
}
\arguments{
\item{sg1_var_name}{Name for subgroup 1 (default: \code{"1"})}

\item{sg2_var_name}{Name for subgroup 2 (default: \code{"2"})}

\item{sg3_var_name}{Name for subgroup 3 (default: \code{"3"})}

\item{sg4_var_name}{Name for subgroup 4 (default: \code{"4"})}

\item{fem_var_name}{Name for Female (default: \code{"Female"})}

\item{mas_var_name}{Name for Male (default: \code{"Male"})}

\item{case_var_name}{Name for Cases (default: \code{"case"})}

\item{control_var_name}{Name for Controls (default: \code{"control"})}
}
\value{
A named vector of hex color codes.
}
\description{
Returns a named vector of colors associated with subgroups, sex, and case-control status.
}
\examples{
subgroup_colors = acr_subgroups_paper_colors()
}
