% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_eval.R
\name{cluster_eval}
\alias{cluster_eval}
\title{Evaluate Clustering with Silhouette Scores}
\usage{
cluster_eval(
  x,
  ndim = 10,
  sillplot_dot_size = 6,
  save_plot = TRUE,
  nclust = NA
)
}
\arguments{
\item{x}{A dissimilarity object of class dissimilarity.}

\item{ndim}{Integer. The maximum number of clusters to consider for silhouette analysis. Default is 10.}

\item{sillplot_dot_size}{Numeric. The size of the points in the silhouette plot. Default is 6.}

\item{save_plot}{Logical. If TRUE, saves the silhouette plot as a JPEG file. Default is TRUE.}

\item{nclust}{Optional. An integer specifying the desired number of clusters. If not provided, the function uses the number of clusters with the highest silhouette width.}
}
\value{
A list containing:
\item{silhouette_plot}{A ggplot2 object for the silhouette plot.}
\item{pam_fit}{The best PAM clustering model based on the optimal number of clusters determined by silhouette width.}
}
\description{
This function performs clustering evaluation by calculating silhouette widths for a specified range of cluster numbers and selecting the optimal cluster number based on the highest average silhouette width. The function can also plot silhouette widths and save the plot.
}
\examples{
# Example usage with a dissimilarity matrix `diss_matrix`
results <- cluster_eval(x = diss_matrix, ndim = 10)
results$silhouette_plot  # Access the silhouette plot
results$pam_fit  # Access the best PAM clustering model

}
