% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ki_colors.R
\name{ki_colors}
\alias{ki_colors}
\title{Retrieves KI Colors}
\usage{
ki_colors(color_palette = "primary", palette_for_printing = FALSE)
}
\arguments{
\item{color_palette}{Color palette names can be set to primary or functional (default: \code{"primary"})}

\item{print_palette}{To display color for printing or web designs (default: \code{"FALSE"})}
}
\value{
A named vector of hex color codes.
}
\description{
Returns a named vector of Karolinska Institutet's palette.
}
\examples{
ki_colors = ki_colors(color_palette = "primary", palette_for_printing = FALSE)
}
