% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lineplot.R
\name{make_aa_lineplot}
\alias{make_aa_lineplot}
\title{Make line plot for autoantibody frequencies per subgroup}
\usage{
make_aa_lineplot(
  x,
  aa_colnames,
  subgroup_column,
  subgroup_colors = distinctColorPalette(k = length(unique(x \%>\%
    pull(subgroup_column)))),
  line_width = 1.5,
  geom_point = 2,
  plot_theme = autoclust::theme_clustering(text_size = 8, plot_title_size = 18),
  ...
)
}
\arguments{
\item{x}{Data frame with autoantibody and subgroup columns.}

\item{aa_colnames}{Character vector of antibody columns.}

\item{subgroup_column}{Name of the column defining subgroups.}

\item{subgroup_colors}{Optional custom colors.}

\item{line_width}{Line thickness.}

\item{geom_point}{Point size.}

\item{plot_theme}{Plot theme (default: autoclust theme).}

\item{...}{Additional arguments passed to ggplot2::labs.}
}
\value{
A ggplot2 object showing antibody frequencies.
}
\description{
Make line plot for autoantibody frequencies per subgroup
}
