% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_most_simmilar_dissimilar.R
\name{plot_most_similar_dissimilar}
\alias{plot_most_similar_dissimilar}
\title{Plot Most Similar and Most Dissimilar Pairs from a Dissimilarity Matrix}
\usage{
plot_most_similar_dissimilar(x, df, id_col)
}
\arguments{
\item{x}{A dissimilarity matrix or an object of class \code{dissimilarity}.}

\item{df}{A dataframe containing individual IDs.}

\item{id_col}{A string specifying the column in \code{df} with individual IDs.}
}
\value{
A \code{ggplotly} object displaying the heatmap of the dissimilarity matrix.
The most similar and most dissimilar pairs are highlighted.
}
\description{
This function visualizes the most similar and most dissimilar pairs of individuals
based on a dissimilarity matrix. It returns an interactive heatmap showing the
dissimilarity values, with the most similar pair highlighted in blue and the most
dissimilar pair highlighted in red.
}
\examples{
# Assuming `dissim_matrix` is your dissimilarity matrix and `id_df` is your ID dataframe
# plot_most_similar_dissimilar(dissim_matrix, id_df, "ID")
}
