% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clustering.R
\name{theme_clustering}
\alias{theme_clustering}
\title{Custom ggplot2 theme for subgroups SLE}
\usage{
theme_clustering(
  text_family = "Roboto",
  text_size = 8,
  text_color = "black",
  plot_title_family = "Lobster Two",
  plot_title_size = 20,
  plot_title_face = "bold",
  plot_title_color = "#2a475e",
  axis_title_size = 12,
  axis_text_size = 10,
  axis_text_color = "black",
  caption_text_color = "black",
  caption_text_size = 8,
  caption_face_style = "bold",
  legend_text_size = 12,
  legend_text_color = "black",
  legend_title_face_style = "bold",
  legend_position = "top"
)
}
\arguments{
\item{text_family}{Font family for general plot text.}

\item{text_size}{Base text size.}

\item{text_color}{Color for general plot text.}

\item{plot_title_family}{Font family for the plot title.}

\item{plot_title_size}{Font size for the plot title.}

\item{plot_title_face}{Font face for the plot title (e.g., \code{"bold"}, \code{"italic"}).}

\item{plot_title_color}{Color for the plot title.}

\item{axis_title_size}{Font size for axis titles.}

\item{axis_text_size}{Font size for axis tick labels.}

\item{axis_text_color}{Color for axis tick labels.}

\item{caption_text_color}{Color of the caption text.}

\item{caption_text_size}{Size of the caption text.}

\item{caption_face_style}{Face style of the caption text (e.g., \code{"bold"}).}

\item{legend_text_size}{Font size for legend text.}

\item{legend_text_color}{Color of legend text.}

\item{legend_title_face_style}{Face style for the legend title (e.g., \code{"bold"}).}

\item{legend_position}{Position of the legend (e.g., \code{"top"}, \code{"bottom"}, \code{"none"}).}
}
\value{
A \code{ggplot2} theme object to be added to a plot.
}
\description{
Returns a \code{ggplot2} theme object with a tailored style optimized for clustering
or publication-quality plots. Allows full customization of font family, size,
color, and positioning for all plot elements.
}
\examples{
library(ggplot2)
ggplot(antibody_dataset, aes(x = Subgroup, y = `HLA PRS 4 SNP`)) +
  geom_boxplot() +
  theme_clustering()

}
