% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dissimilarity_matrix.R
\name{compute_dissimilarity_matrix}
\alias{compute_dissimilarity_matrix}
\title{Compute dissimilarity matrices}
\usage{
compute_dissimilarity_matrix(
  x,
  metric = c("euclidean", "manhattan", "gower"),
  stand = FALSE,
  return_dis = FALSE,
  plot_dissim = TRUE,
  heatmap_title = paste0(metric, " Dissimilarity Matrix"),
  heatmap_axis_text_size = 4,
  plot_title_size = 20,
  save_plots = FALSE,
  plot_outpath = "."
)
}
\arguments{
\item{x}{numeric matrix or data frame (n x p). Dissimilarities are computed between the rows of x. Dataframe must contain headers.
For further details, see help(daisy).}

\item{metric}{Metric to be used to compute matrix. Possible options include "gower" (default), "euclidean", or "manhattan".}

\item{stand}{if TRUE, measurements in x are standardized before calculating dissimilarities. For further details, see help(daisy).}

\item{return_dis}{if TRUE function returns dissimilarity object rather than matrix.}

\item{plot_dissim}{if TRUE, plots the dissimilarity matrix as a heatmap and shows hierarchical clustering on the plot.}

\item{heatmap_title}{Title for the heatmap plot.}

\item{heatmap_axis_text_size}{Heatmap x-y axis text size.}

\item{plot_title_size}{Font size for the plot title.}

\item{save_plots}{If TRUE plots produced are saved in the working directory.}

\item{plot_outpath}{Output pathway for plots. If not specified and save_plots is TRUE they are saved in the current directory.}
}
\value{
A list containing the dissimilarity matrix, optionally the heatmap plot if specified.
}
\description{
Wrapper function (daisy) for computing dissimilarity matrices from matrix or datasets of continuous, discrete, or categorical variables.
The function also creates a heatmap using the dissimilarity matrix if specified.
}
\examples{
# Example of how to use the function
result <- compute_dissimilarity_matrix(x = autoant_dataset, metric = "gower")
result$dissimilarity_matrix  # Access the matrix
result$heatmap_plot  # Access the plot (if plot_dissim = TRUE)

}
