% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_dissimilarity_matrix.R
\name{compute_dissimilarity_matrix}
\alias{compute_dissimilarity_matrix}
\title{Compute dissimilarity matrices}
\usage{
compute_dissimilarity_matrix(
  x,
  metric = c("euclidean", "manhattan", "gower"),
  stand = FALSE,
  plot_dissim = TRUE
)
}
\arguments{
\item{x}{numeric matrix or dataframe n x p. Dissimilarities are computed between the rows of x. Dataframe must contain headers. For further details see help(daisy).}

\item{metric}{Metric to be used to compute matrix. Possible options include gower(default), euclidean or manhattan.}

\item{stand}{if TRUE, measurements in x are standarized before calculating dissimilarities. For further details see help(daisy).}

\item{plot_dissim}{Plots the dissimilarity matrix as a heatmap and shows hirarchical clustering on plot.}

\item{row_names}{if TRUE, df/matrix contains row names and will be removed by function.}
}
\value{
Description of what the function returns
}
\description{
Wrapper function (daisy) for to compute dissimilarity matrices from matrix or datasets of continous, discrete or categorical variables.
}
\examples{
# Example of how to use the function
compute_dissimilarity_matrix(x = autoant_dataset, metric = "gower")

}
